#303: prepared integration tests setup
