+++ src/test/java/com/exasol/adapter/dialects/postgresql/PostgreSQLSqlDialectIT.java
@@ -13,13 +13,10 @@ import java.util.Collections;
 import java.util.Optional;
 import java.util.concurrent.TimeoutException;
 
-import javax.sql.DataSource;
-
 import org.hamcrest.MatcherAssert;
 import org.junit.jupiter.api.*;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.testcontainers.containers.JdbcDatabaseContainer;
 import org.testcontainers.containers.PostgreSQLContainer;
 import org.testcontainers.containers.output.Slf4jLogConsumer;
 import org.testcontainers.junit.jupiter.Container;
@@ -30,8 +27,6 @@ import com.exasol.bucketfs.BucketAccessException;
 import com.exasol.containers.ExasolContainer;
 import com.exasol.containers.ExasolContainerConstants;
 import com.exasol.jdbc.DataException;
-import com.zaxxer.hikari.HikariConfig;
-import com.zaxxer.hikari.HikariDataSource;
 
 import utils.IntegrationTestSetupManager;
 
@@ -40,7 +35,7 @@ import utils.IntegrationTestSetupManager;
 class PostgreSQLSqlDialectIT {
     private static final Logger LOGGER = LoggerFactory.getLogger(PostgreSQLSqlDialectIT.class);
     private static final String CONNECTION_POSTGRES_JDBC = "CONNECTION_POSTGRES_JDBC";
-    private static final String SCHEMA_POSTGRES_TEST = "schema_postgres_test";
+    private static final String SCHEMA_POSTGRES = "schema_postgres";
     private static final String SCHEMA_POSTGRES_UPPERCASE_TABLE = "schema_postgres_upper";
     private static final String TABLE_POSTGRES_SIMPLE = "table_postgres_simple";
     private static final String TABLE_POSTGRES_MIXED_CASE = "Table_Postgres_Mixed_Case";
@@ -51,7 +46,6 @@ class PostgreSQLSqlDialectIT {
     private static final String VIRTUAL_SCHEMA_POSTGRES_PRESERVE_ORIGINAL_CASE = "VIRTUAL_SCHEMA_POSTGRES_PRESERVE_ORIGINAL_CASE";
     private static final String QUALIFIED_TABLE_JOIN_NAME_1 = VIRTUAL_SCHEMA_POSTGRES + "." + TABLE_JOIN_1;
     private static final String QUALIFIED_TABLE_JOIN_NAME_2 = VIRTUAL_SCHEMA_POSTGRES + "." + TABLE_JOIN_2;
-    private static final String DOCKER_IP_ADDRESS = "172.17.0.1";
     private static final int POSTGRES_PORT = 5432;
 
     @Container
@@ -71,53 +65,38 @@ class PostgreSQLSqlDialectIT {
         final Connection exasolConnection = exasolContainer.createConnectionForUser(exasolContainer.getUsername(),
                 exasolContainer.getPassword());
         statementExasol = exasolConnection.createStatement();
-        final Statement statementPostgres = getStatement(postgresqlContainer);
-        integrationTestSetupManager.createTestSchema(statementExasol, SCHEMA_EXASOL_TEST);
-        integrationTestSetupManager.createTestSchema(statementPostgres, SCHEMA_POSTGRES_TEST);
+        final Statement statementPostgres = integrationTestSetupManager.getStatement(postgresqlContainer);
+        integrationTestSetupManager.createTestSchema(statementExasol, SCHEMA_EXASOL);
+        integrationTestSetupManager.createTestSchema(statementPostgres, SCHEMA_POSTGRES);
         integrationTestSetupManager.createTestSchema(statementPostgres, SCHEMA_POSTGRES_UPPERCASE_TABLE);
         createPostgresTestTableSimple(statementPostgres);
         createPostgresTestTableAllDataTypes(statementPostgres);
         createPostgresTestTableMixedCase(statementPostgres);
         createPostgresTestTableLowerCase(statementPostgres);
-        integrationTestSetupManager.createTestTablesForJoinTests(statementPostgres, SCHEMA_POSTGRES_TEST, TABLE_JOIN_1,
+        integrationTestSetupManager.createTestTablesForJoinTests(statementPostgres, SCHEMA_POSTGRES, TABLE_JOIN_1,
                 TABLE_JOIN_2);
         final String connectionString = "jdbc:postgresql://" + DOCKER_IP_ADDRESS + ":"
                 + postgresqlContainer.getMappedPort(POSTGRES_PORT) + "/" + postgresqlContainer.getDatabaseName();
-        System.out.println(connectionString);
         integrationTestSetupManager.createConnection(statementExasol, CONNECTION_POSTGRES_JDBC, connectionString,
                 postgresqlContainer.getUsername(), postgresqlContainer.getPassword());
-        integrationTestSetupManager.createAdapterScript(statementExasol,
-                SCHEMA_EXASOL_TEST + "." + ADAPTER_SCRIPT_EXASOL, VIRTUAL_SCHEMAS_JAR_NAME_AND_VERSION,
+        integrationTestSetupManager.createAdapterScript(statementExasol, SCHEMA_EXASOL + "." + ADAPTER_SCRIPT_EXASOL,
+                VIRTUAL_SCHEMAS_JAR_NAME_AND_VERSION,
                 Optional.of("%jar /buckets/bfsdefault/default/" + POSTGRES_DRIVER_NAME_AND_VERSION + ";\n"));
-        createVirtualSchema(VIRTUAL_SCHEMA_POSTGRES, SCHEMA_POSTGRES_TEST, Optional.empty());
+        createVirtualSchema(VIRTUAL_SCHEMA_POSTGRES, SCHEMA_POSTGRES, Optional.empty());
         createVirtualSchema(VIRTUAL_SCHEMA_POSTGRES_UPPERCASE_TABLE, SCHEMA_POSTGRES_UPPERCASE_TABLE,
                 Optional.of("ignore_errors='POSTGRESQL_UPPERCASE_TABLES'"));
         createVirtualSchema(VIRTUAL_SCHEMA_POSTGRES_PRESERVE_ORIGINAL_CASE, SCHEMA_POSTGRES_UPPERCASE_TABLE,
                 Optional.of("POSTGRESQL_IDENTIFIER_MAPPING = 'PRESERVE_ORIGINAL_CASE'"));
     }
 
-    private static Statement getStatement(final JdbcDatabaseContainer container) throws SQLException {
-        final DataSource dataSource = getDataSource(container);
-        return dataSource.getConnection().createStatement();
-    }
-
-    private static DataSource getDataSource(final JdbcDatabaseContainer container) {
-        final HikariConfig hikariConfig = new HikariConfig();
-        hikariConfig.setJdbcUrl(container.getJdbcUrl());
-        hikariConfig.setUsername(container.getUsername());
-        hikariConfig.setPassword(container.getPassword());
-        hikariConfig.setDriverClassName(container.getDriverClassName());
-        return new HikariDataSource(hikariConfig);
-    }
-
     private static void createPostgresTestTableSimple(final Statement statementPostgres) throws SQLException {
-        final String qualifiedTableName = SCHEMA_POSTGRES_TEST + "." + TABLE_POSTGRES_SIMPLE;
+        final String qualifiedTableName = SCHEMA_POSTGRES + "." + TABLE_POSTGRES_SIMPLE;
         statementPostgres.execute("CREATE TABLE " + qualifiedTableName + " (x INT)");
         statementPostgres.execute("INSERT INTO " + qualifiedTableName + " VALUES (1)");
     }
 
     private static void createPostgresTestTableAllDataTypes(final Statement statementPostgres) throws SQLException {
-        final String qualifiedTableName = SCHEMA_POSTGRES_TEST + "." + TABLE_POSTGRES_ALL_DATA_TYPES;
+        final String qualifiedTableName = SCHEMA_POSTGRES + "." + TABLE_POSTGRES_ALL_DATA_TYPES;
         statementPostgres.execute("CREATE TABLE " + qualifiedTableName + " (" //
                 + "myBigint BIGINT,	" //
                 + "myBigserial BIGSERIAL, " //
@@ -213,7 +192,7 @@ class PostgreSQLSqlDialectIT {
         final StringBuilder builder = new StringBuilder();
         builder.append("CREATE VIRTUAL SCHEMA ");
         builder.append(virtualSchemaName);
-        builder.append(" USING " + SCHEMA_EXASOL_TEST + "." + ADAPTER_SCRIPT_EXASOL + " WITH ");
+        builder.append(" USING " + SCHEMA_EXASOL + "." + ADAPTER_SCRIPT_EXASOL + " WITH ");
         builder.append("SQL_DIALECT     = 'POSTGRESQL' ");
         builder.append("CATALOG_NAME     = '" + postgresqlContainer.getDatabaseName() + "' ");
         builder.append("CONNECTION_NAME = '" + CONNECTION_POSTGRES_JDBC + "' ");
@@ -226,7 +205,7 @@ class PostgreSQLSqlDialectIT {
 
     @Test
     void testSelectSingleColumn() throws SQLException {
-        final String qualifiedExpectedTableName = SCHEMA_EXASOL_TEST + ".TABLE_POSTGRES_SIMPLE_EXPECTED";
+        final String qualifiedExpectedTableName = SCHEMA_EXASOL + ".TABLE_POSTGRES_SIMPLE_EXPECTED";
         statementExasol.execute("CREATE OR REPLACE TABLE " + qualifiedExpectedTableName + " (x INT)");
         statementExasol.execute("INSERT INTO " + qualifiedExpectedTableName + " VALUES (1)");
         final ResultSet expectedResultSet = statementExasol.executeQuery("SELECT * FROM " + qualifiedExpectedTableName);
@@ -241,7 +220,7 @@ class PostgreSQLSqlDialectIT {
         @Test
         void testInnerJoin() throws SQLException {
             final ResultSet expected = integrationTestSetupManager.getSelectAllFromJoinExpectedTable(statementExasol,
-                    SCHEMA_EXASOL_TEST, "(x INT, y VARCHAR(100), a INT, b VARCHAR(100))", "VALUES(2,'bbb', 2,'bbb')");
+                    SCHEMA_EXASOL, "(x INT, y VARCHAR(100), a INT, b VARCHAR(100))", "VALUES(2,'bbb', 2,'bbb')");
             final ResultSet actual = statementExasol.executeQuery("SELECT * FROM " + QUALIFIED_TABLE_JOIN_NAME_1
                     + " a INNER JOIN  " + QUALIFIED_TABLE_JOIN_NAME_2 + " b ON a.x=b.x");
             MatcherAssert.assertThat(actual, matchesResultSet(expected));
@@ -250,7 +229,7 @@ class PostgreSQLSqlDialectIT {
         @Test
         void testInnerJoinWithProjection() throws SQLException {
             final ResultSet expected = integrationTestSetupManager.getSelectAllFromJoinExpectedTable(statementExasol,
-                    SCHEMA_EXASOL_TEST, "(y VARCHAR(100))", " VALUES('bbbbbb')");
+                    SCHEMA_EXASOL, "(y VARCHAR(100))", " VALUES('bbbbbb')");
             final ResultSet actual = statementExasol.executeQuery("SELECT b.y || " + QUALIFIED_TABLE_JOIN_NAME_1
                     + ".y FROM " + QUALIFIED_TABLE_JOIN_NAME_1 + " INNER JOIN  " + QUALIFIED_TABLE_JOIN_NAME_2
                     + " b ON " + QUALIFIED_TABLE_JOIN_NAME_1 + ".x=b.x");
@@ -260,8 +239,7 @@ class PostgreSQLSqlDialectIT {
         @Test
         void testLeftJoin() throws SQLException {
             final ResultSet expected = integrationTestSetupManager.getSelectAllFromJoinExpectedTable(statementExasol,
-                    SCHEMA_EXASOL_TEST, "(x INT, y VARCHAR(100), a INT, b VARCHAR(100))",
-                    "VALUES(1, 'aaa', null, null), " //
+                    SCHEMA_EXASOL, "(x INT, y VARCHAR(100), a INT, b VARCHAR(100))", "VALUES(1, 'aaa', null, null), " //
                             + "(2, 'bbb', 2, 'bbb')");
             final ResultSet actual = statementExasol.executeQuery("SELECT * FROM " + QUALIFIED_TABLE_JOIN_NAME_1
                     + " a LEFT OUTER JOIN  " + QUALIFIED_TABLE_JOIN_NAME_2 + " b ON a.x=b.x ORDER BY a.x");
@@ -271,7 +249,7 @@ class PostgreSQLSqlDialectIT {
         @Test
         void testRightJoin() throws SQLException {
             final ResultSet expected = integrationTestSetupManager.getSelectAllFromJoinExpectedTable(statementExasol,
-                    SCHEMA_EXASOL_TEST, "(x INT, y VARCHAR(100), a INT, b VARCHAR(100))", "VALUES(2, 'bbb', 2, 'bbb'), " //
+                    SCHEMA_EXASOL, "(x INT, y VARCHAR(100), a INT, b VARCHAR(100))", "VALUES(2, 'bbb', 2, 'bbb'), " //
                             + "(null, null, 3, 'ccc')");
             final ResultSet actual = statementExasol.executeQuery("SELECT * FROM " + QUALIFIED_TABLE_JOIN_NAME_1
                     + " a RIGHT OUTER JOIN  " + QUALIFIED_TABLE_JOIN_NAME_2 + " b ON a.x=b.x ORDER BY a.x");
@@ -281,8 +259,7 @@ class PostgreSQLSqlDialectIT {
         @Test
         void testFullOuterJoin() throws SQLException {
             final ResultSet expected = integrationTestSetupManager.getSelectAllFromJoinExpectedTable(statementExasol,
-                    SCHEMA_EXASOL_TEST, "(x INT, y VARCHAR(100), a INT, b VARCHAR(100))",
-                    "VALUES(1, 'aaa', null, null), " //
+                    SCHEMA_EXASOL, "(x INT, y VARCHAR(100), a INT, b VARCHAR(100))", "VALUES(1, 'aaa', null, null), " //
                             + "(2, 'bbb', 2, 'bbb'), " //
                             + "(null, null, 3, 'ccc')");
             final ResultSet actual = statementExasol.executeQuery("SELECT * FROM " + QUALIFIED_TABLE_JOIN_NAME_1
@@ -293,7 +270,7 @@ class PostgreSQLSqlDialectIT {
         @Test
         void testRightJoinWithComplexCondition() throws SQLException {
             final ResultSet expected = integrationTestSetupManager.getSelectAllFromJoinExpectedTable(statementExasol,
-                    SCHEMA_EXASOL_TEST, "(x INT, y VARCHAR(100), a INT, b VARCHAR(100))", "VALUES(2, 'bbb', 2, 'bbb'), " //
+                    SCHEMA_EXASOL, "(x INT, y VARCHAR(100), a INT, b VARCHAR(100))", "VALUES(2, 'bbb', 2, 'bbb'), " //
                             + "(null, null, 3, 'ccc')");
             final ResultSet actual = statementExasol.executeQuery("SELECT * FROM " + QUALIFIED_TABLE_JOIN_NAME_1
                     + " a RIGHT OUTER JOIN  " + QUALIFIED_TABLE_JOIN_NAME_2 + " b ON a.x||a.y=b.x||b.y ORDER BY a.x");
@@ -303,8 +280,7 @@ class PostgreSQLSqlDialectIT {
         @Test
         void testFullOuterJoinWithComplexCondition() throws SQLException {
             final ResultSet expected = integrationTestSetupManager.getSelectAllFromJoinExpectedTable(statementExasol,
-                    SCHEMA_EXASOL_TEST, "(x INT, y VARCHAR(100), a INT, b VARCHAR(100))",
-                    "VALUES(1, 'aaa', null, null), " //
+                    SCHEMA_EXASOL, "(x INT, y VARCHAR(100), a INT, b VARCHAR(100))", "VALUES(1, 'aaa', null, null), " //
                             + "(2, 'bbb', 2, 'bbb'), " //
                             + "(null, null, 3, 'ccc')");
             final ResultSet actual = statementExasol.executeQuery("SELECT * FROM " + QUALIFIED_TABLE_JOIN_NAME_1
@@ -327,17 +303,17 @@ class PostgreSQLSqlDialectIT {
         @Test
         void testQueryUpperCaseTableQuotedThrowsException() {
             final Exception exception = assertThrows(SQLException.class, () -> statementExasol
-                    .execute("SELECT x FROM  " + SCHEMA_EXASOL_TEST + ".\"" + TABLE_POSTGRES_MIXED_CASE + "\""));
-            assertThat(exception.getMessage(), containsString(
-                    "object \"" + SCHEMA_EXASOL_TEST + "\".\"" + TABLE_POSTGRES_MIXED_CASE + "\" not found"));
+                    .execute("SELECT x FROM  " + SCHEMA_EXASOL + ".\"" + TABLE_POSTGRES_MIXED_CASE + "\""));
+            assertThat(exception.getMessage(),
+                    containsString("object \"" + SCHEMA_EXASOL + "\".\"" + TABLE_POSTGRES_MIXED_CASE + "\" not found"));
         }
 
         @Test
         void testQueryUpperCaseTableThrowsException() {
-            final Exception exception = assertThrows(SQLException.class, () -> statementExasol
-                    .execute("SELECT x FROM  " + SCHEMA_EXASOL_TEST + "." + TABLE_POSTGRES_MIXED_CASE));
+            final Exception exception = assertThrows(SQLException.class,
+                    () -> statementExasol.execute("SELECT x FROM  " + SCHEMA_EXASOL + "." + TABLE_POSTGRES_MIXED_CASE));
             assertThat(exception.getMessage(), containsString(
-                    "object " + SCHEMA_EXASOL_TEST + "." + TABLE_POSTGRES_MIXED_CASE.toUpperCase() + " not found"));
+                    "object " + SCHEMA_EXASOL + "." + TABLE_POSTGRES_MIXED_CASE.toUpperCase() + " not found"));
         }
 
         @Test
@@ -413,7 +389,7 @@ class PostgreSQLSqlDialectIT {
 
         private void assertSingleValue(final String columnName, final String expectedColumnType,
                 final String expectedValue) throws SQLException {
-            final String qualifiedExpectedTableName = SCHEMA_EXASOL_TEST + "." + "EXPECTED";
+            final String qualifiedExpectedTableName = SCHEMA_EXASOL + "." + "EXPECTED";
             statementExasol.execute(
                     "CREATE OR REPLACE TABLE " + qualifiedExpectedTableName + "(x " + expectedColumnType + ")");
             statementExasol.execute("INSERT INTO " + qualifiedExpectedTableName + " VALUES(" + expectedValue + ")");
