+++ src/test/resources/integration/driver/oracle/settings.cfg
@@ -1,0 +1,6 @@
+DRIVERNAME=ORACLE
+JAR=ojdbc10.jar
+DRIVERMAIN=oracle.jdbc.driver.OracleDriver
+PREFIX=jdbc:oracle:thin:
+FETCHSIZE=100000
+INSERTSIZE=-1
